gs_terraform/src/fstore-mg-cloud-function-v2/requirements.txt
functions-framework==3.*
python-dotenv==1.*
google-cloud-firestore
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-bigquery


